Mandatory part done

+++ ft_putuns.c
@@ -1,9 +1,9 @@
 #include "libft.h"
-#include "libprintf.h"
+#include "ft_printf.h"
 
-void    ft_putuns(unsigned long long n, int *res)
+void	ft_putuns(unsigned int n, int *res)
 {
-    if (n >= 10)
+	if (n >= 10)
 	{
 		ft_putuns(n / 10, res);
 		n = n % 10;
